In the baseline + committe models the variance of the committee is defined as (see also Eq. 11 in https://doi.org/10.1063/5.0036522 ):
\sigma^2 = 1/(M-1) \sum_{i=0}^M (V_i - \bar{V})^2

where the sum runs over the M models of the committee, V_i is the potential of i-th model and \bar{V} the mean potential of the committee.
the formula is correctly implemented in the variables var_pot , where the sum is divided by (M-1).

for the derivative the term uncertain_frc and uncertain_vir contained a 1/M which created a small inconsistence.
This commit fix this issue.