A bunch of profiling-driven optimization for i-PI
This PR implements some pretty extensive changes to the core infrastructure, aimed at reducing the overhead of i-PI in low-cost-simulations scenarios (benchmarks are NVE trajectories with dummy drivers that return zeros, or that do not compute anything at all). 

The main (most invasive and potentially most dangerous) changes are:
* switching almost everywhere (softexit being the only exception since it's created only once) from using `threading` to make new threads to `concurrent.futures.ThreadPoolExecutor` that seems to have lower overhead for launching a new worker
* changing the way the "syntactic sugar" around depend objects is implemented. rather than hijacking __getattribute___ for the parent class (that means a small overhead is added to each member access) I used @property-style accessors so the overhead is only for the depend object members. This incidentally also simplifies the usage pattern, because you can access the "real" depend object with parent._object.

Took this opportunity also to add a `profiling` folder and to document a tiny bit the depend_value machinery, so new developers have somewhere to start. 
---------

Co-authored-by: Michele Ceriotti <michele.ceriotti@gmail.com>
Co-authored-by: Venkat Kapil <venkat.kapil@gmail.com>
Co-authored-by: Mariana Rossi <rossi@fhi-berlin.mpg.de>