Fix aspects of interface between i-Pi and PLUMED that are not working (#340)
* Fix aspects of interface between i-Pi and PLUMED that are not working

There are two problems with the interface between i-Pi and PLUMED at the moment:

(1) The timestep is set incorrectly.  It is actually not possible to set the timestep correctly.  I have thus deleted the command to set the PLUMED timestep here.
(2) The cell appears to be passed incorrect.  If you pass a deep copy of the NumPy array containing the cell as I do here I think the problem is fixed.

* Ensured that PLUMED always interprets the timestep for i-Pi as one.

---------

Co-authored-by: Gareth Aneurin Tribello <garethtribello@gareths-mbp.mp.qub.ac.uk>
Co-authored-by: Gareth Aneurin Tribello <garethtribello@Gareths-MacBook-Pro.local>
Co-authored-by: Michele Ceriotti <michele.ceriotti@gmail.com>