A "direct" FF using the same PES interfaces as the python driver (#390)
This is a major restructuring of the python driver and the forcefield block. 
All PES drivers have been moved to an ipi/pes folder, and their API streamlined
with the possibility of providing both lists of arguments and named arguments.
A new FFDirect forcefield makes it possible to call the PES from i-PI directly, avoiding
the client/server setup that many find confusing. 
We converted most PES to the new interfaces, but several are possibly broken;
these now trigger a warning so that users can check and possibly fix them.


---------

Co-authored-by: Yair Litman <yairlitman@gmail.com>
Co-authored-by: Venkat Kapil <venkat.kapil@gmail.com>