Major update to how the plumed module collects hills work
This avoids a lot of unnecessary computations, and one can switch off
the calculation of total work to save even more.
Note that this change makes calculations somewhat less robust, as the
assumption mtd_update is called right after having updated the bias
i-pi-side is even more important.
However this was an assumption all along, so overall this is just being
more consequential about it.