Bugfix/get temp (#417)
get_temp has been giving weird results, and confusing people, for a long time. 
This finally  tackles the problem heads-on, removing a weird workaround to split
the COM kinetic energy when computing temperature for a subset of the atoms,
and issuing warning when the temperature is not well-defined (e.g. when looking
at beads with a non-constant NM mass matrix). All regtest are also updated, and 
a few have been added.

---------

Co-authored-by: Wei Fang <wei_fang@fudan.edu.cn>
Co-authored-by: Michele Ceriotti <ceriottm@users.noreply.github.com>
Co-authored-by: Michele Ceriotti <michele.ceriotti@gmail.com>